Move FFT and power-analysis to signal_processing
