+++ scos_actions/actions/tests/test_acquire_single_freq_fft.py
@@ -1,8 +1,12 @@
-from scos_actions.actions.fft import M4sDetector
-from scos_actions.actions.interfaces.signals import measurement_action_completed
-from scos_actions.actions.tests.utils import SENSOR_DEFINITION, check_metadata_fields
-from scos_actions.discover import test_actions as actions
+from scos_actions.actions.interfaces.signals import (
+    measurement_action_completed
+)
+from scos_actions.actions.tests.utils import (
+    SENSOR_DEFINITION,
+    check_metadata_fields
+)
 from scos_actions.capabilities import capabilities
+from scos_actions.discover import test_actions as actions
 
 SINGLE_FREQUENCY_FFT_ACQUISITION = {
     "name": "test_acq",
@@ -12,6 +16,7 @@ SINGLE_FREQUENCY_FFT_ACQUISITION = {
     "action": "test_single_frequency_m4s_action",
 }
 capabilities['sensor'] = SENSOR_DEFINITION
+
 
 def test_detector():
     _data = None
